+++ src/components/form/field/index.js
@@ -27,6 +27,10 @@ var _radioButtonSet = require("./components/radio-button-set");
 
 var _select = require("./components/select");
 
+var _checkbox = require("./components/checkbox");
+
+var _password = require("./components/password");
+
 var Field =
 /*#__PURE__*/
 function (_Component) {
@@ -60,4 +64,6 @@ Field.Label = _label.FieldLabel;
 Field.Input = _input.FieldInput;
 Field.RadioButtonSet = _radioButtonSet.FieldRadioButtonSet;
 Field.Select = _select.FieldSelect;
+Field.Checkbox = _checkbox.FieldCheckbox;
 Field.ErrorMessage = _error.FieldError;
+Field.Password = _password.FieldPassword;