+++ CHANGELOG.md
@@ -8,6 +8,36 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Breaking Change: needs React 18
 - Feature: New `StickyContainer` component.
 
+## [1.25.0] - 2019-03-21
+
+Feature:
+- Add `lightRed` (`#ffebe0`) to `COLORS` set.
+- Add `unit` prop to `Field.Input`.
+
+## [1.24.0] - 2019-03-21
+
+Features:
+- Add `link` props to `Field.Label`
+- Add `Field.Password` component
+
+## [1.23.0] - 2019-03-20
+
+Features:
+- Add `Field.Checkbox`.
+- Export `FLUID`, `DEFAULT`, `TINY` and `BIG` css from `Button`.
+
+Fix:
+- Update selected option style to `Select` component.
+
+## [1.22.0] - 2019-03-18
+
+Features:
+- Add `customSize` props to `HorizontalStroke`.
+- Add `HorizontalStroke` size CSS: `TINY`, `DEFAULT`, `BIG` and `HUGE`.
+- Add `IconList`.
+- Add empty `Prev` and `Next` container to `NavTabCarousel` when
+  `NavTabCarousel.Prev` or `NavTabCarousel.Next` are empty.
+
 ## [1.21.0] - 2019-03-11
 
 Features: